[Agent] Organize bootstrapper stages
